[spr] initial version
Created using spr 1.3.7-wip
