+++ llvm/lib/Remarks/BitstreamRemarkParser.cpp
@@ -11,7 +11,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "llvm/Remarks/BitstreamRemarkParser.h"
 #include "BitstreamRemarkParser.h"
 #include "llvm/Remarks/Remark.h"
 #include "llvm/Support/MemoryBuffer.h"
